+++ Campaign/extensions/campaign-module/database/migrations/2025_11_13_132323_increase_target_url_length_in_banners_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class IncreaseTargetUrlLengthInBannersTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('banners', function (Blueprint $table) {
+            $table->string('target_url', 768)
+                ->nullable()
+                ->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('banners', function (Blueprint $table) {
+
+        });
+    }
+}